Fluxes: Partially restore heterodyne frequency information.
When the scuba.dat was deprecated and scuba2.dat used, the
frequencies for heterodyne observations were removed.  This change
restores 1300 and adds 868 and 434 micrometre wavelengths
(234.5, 345.8, 691.5 GHz) for ACSIS-era heterodyne instruments.
For DAS spectra the assumed beamsizes and flux densities were
interpolated in frequency between those of the old SCUBA filters.