smurf: Only use ringing filter if AST model was used
If a ringing filter is used on a pre-iteration (i.e. an iteration
specified by AST.SKIP), then it can produce a hole in the map around very
bright sources. This means areas around bright sources are not flagged as
sources in the FLT mask on the next iteration, meaning they are not
excluded from the FLT model and so cause ringing that is never removed.