smurf: Ensure JSA tiles have a defined telescope position in the WCS
The ObsLat, ObsLon and ObsAlt attributes are now set to the position of
JCMT as defined by PAL. I initially copied the position from an input
smfHead, rather than just using the PAL position, until I realised that
the position needs be defined consistently whether or not a smfHead is
available. For instance, no raw data is available when determining a list
of tiles that touch an arbitrary sky region.