smurf: Ensure ObsLat/ObsLon are equal in SkyFrame and SpecFrame
The values stored by makemap in the output SkyFrame are copied to the
specframe. But the SkyFrame values returned by AST are rounded to the
nearest 0.01 arc-second, so the values stored in the SpecFrame end up as
rounded versions of those stored in the SkyFrame. This changes rounds
each value to 0.01 arc-sec (about 30cm on the earth's surface) before
storing them in the skyframe, so both frames end up with the same value.