docs: Move the figures in Section 6 of SC/21.
The aim is to get them to appear in their own sections, with suitable
magnfication.  The ones in displaying mask won't scale to requested
height in the html, so keep with the one \myfig for each, rather than
separate scalings for PDF and HTML.  The positions and location
directives  may change if the document layout is otherwise adjusted.

Used \myfig macro for some recent graphics.

Tidy a caption or two.