docs: Try a different \gaiathing definition in SC/21.
To match the GAIA style better make the gaiathing a bold  san serif
fount.  It's looks fine in the PDF, but the html it's bold italic roman.
Still on one page that had too much bold from a TIP as well as several
GAIA things, the italics delineate the two.  We shall have to see if
Holly approves this change.