docs: Coloured quotes for example commands and output in SC/21.
Holly's change didn't work well in HTML, where early on the quotes
switched to black and the main text was in MidnighrtBlue.  I tried
adding a \color{Black} to restore the normal black text, but that only
succeeding in making the examples green.  Until the HTML aspect is
robust I have added a \renewcommand for HTML to retain the normal
myquote environment.  I would have bracketed the original definition
with htmlonly/latexonly but the point where newenviroments are supposed
to go comes before those are defined.