docs: Strive for consistency in titles in SC/21.
Standardised the capitalisation styles: uppercase for each non-trivial
word in chapters, but only capitals for the first word (except for
names) in the sections and appendices.

Make the section and chapter names in links match the actual titles.

Apply \texttt and \drrecipe wrapper to the file and recipes names
respectively in titles.