docs: Move the section titlesec until the main body of SC/21.
This is to prevent the chapter-style section headings to be used for
the List of Figures and Contents pages, which included the section
number, i.e. 0.  Make the title take the form "Chapter 1: ".

Also remove an trial colour, now unused, for the blue section banners.
Add some commentary so next time we'll have an idea what is being done
and why.