docs: Use two smaller graphics files in SC/21.
The size scaling does not work for some of the PNGs in HTML version
(hence not using \htmlonly for those).  Since some of the recent additions
are too big to fit on a typical browser window, reduce them.  We did
have smaller versions last time around...

There is still one large graphic remaining---the STC region outlines---but
the outlines are broken in the smaller graphic.