docs: Make chapter headings for LaTeX only in SC/21.
This generates headings of the following form.

Chapter <section number>n
Title <title>

This works fine for the PDF, but the Latex2HTML appears not to be
respecting %\begin{latexonly} ... %\end{latexonly} construct.  As a
consequence the HTML has two new sections before the Introduction with
"<: titles and parts of the titlesec defintions.  There must surely be
a way to prevent the titlesec definitions from being accessed for the
HMTL build (given that LaTeX2HTML does not support this package).  If
we have to live with one or the other, we are going with the PDF.