CONVERT: Fix bug preventing creation of some primitive object in extensions.
This relates to importing FITS files that were once converted from an
NDF.  Primitive components at the top level of the NDF were being
propagated correctly, but not ones lower down, such as for a FITS
airlock of an NDF that itself is an extension.  Such a case has arisen
recently for an NELP extension, presumably created by a SCUBA-2/PICARD
recipe.