CONVERT: Access the correct locator index for a top-level primitive extension.
This used to work until a recent change (2093fab26a02) to handle the
same type of component, but buried with an NDF extension.  After the
recent change the latter worked but it needed an index subtraction,
whereas the top-level primitive component is already in SXLOC(1).
Not having an example to hand of this sort of component, other than
the FITS airlock dealt with independently. I fooled myself that it was
working.  The fix was simple.