docs: Add earlier versions of docs not previously in repo history
These are documents found on the JAC VAXes but which have newer
variants already present in the repository. Many of these documents
are useful historically because they refer to VMS versions of
applications.

I am willing to concede that I might not have put them in the
best places in the tree. It may be that they should sit with the
applications/library directories but I didn't want to confuse those
directories by including old documents.

(there are inconsistencies in that SUN/41 has two versions in the
IRCAMDR source tree and previously I had put the SSN/64 in obsolete
and the main ssn tree).