smurf: Do not recalculate BOLO frame
The call to sc2ast_createwcs in smf_create_bolfile was a bit unnecesary
given that the supplied time-series WCS in the reference smfData contains
the information. At the time I was too lazy to extract it but now that
ATL includes atlFrameSetSplit() the excuses are not warranted. Also
now means that we don't have a SCUBA-2 routine in the middle of
an allegedly instrument-agnostic map-maker.