ndg: Prevent HDS indices causing errors after CONVERT is started
After CONVERT is started, NDG needs to check for file types that include
dots (e.g. "fits.gz"). To do this it repeatedly checks each dot-delimited
trailing field in the file name. But if the name contains an HDS cell
index (e.g. ".more.cupid.clumps(2).model") then wordexp reports an error
when it is passed a template include the parentheses (e.g. "(2)" ).

To prevent this, check for the error and annul it. In other words, treat
this error exactly as if it was a simple case of "no matching files
found".