extractor: update the version of SExtractor to 2.19.5.
This is just an attempt to update the code base and keep the existing
abilities. New features such as model fitting, PSF and galaxy, have
not been looked at and are configured out of the build (this avoids
having to look at all this, and also include ATLAS as a dependency,
externally that would be embarrasing, and internally hard to justify
and build). This could be returned to if there was any interest, but
that is probably best left to vanilla builds of SExtractor and PSFEx.