hds: Use public API in Fortran wrapper
Use the public datImportFloc routine for the HDS Fortran wrapper
just like any user of the HDS API would from C for Fortran.
This allows the same code to be tested inside HDS as outside.

DAT_WHERE is anomalous in that it has no public C interface
and the wrapper requires knowledge of internals such as INT_BIG
and dat1emsSetBigi. If DAT_WHERE was removed from the HDS API
the entire Fortran wrapper would be implemented using public
APIs.