ast: Allow wrapping of lon/lat values to be disabled in WcsMap
Provide a class variable ("WcsWrap") that allow the wrapping of
lon/lat into their normal ranges to be disabled prior to
transforming them into (x,y) values. The flag can be set and got
using the class (i.e. non-virtual) functions astGetWcsWrap and
astSetWcsWrap. The default is to do wrapping. It may be
necessary to disable wrapping for projections that have complex
definition of the "normal" range of longitude. For instance, the
normal range of longitude for an all-sky HPX projection depends
on the latitude so that wrap-around happens on a zig-zag line
rather than a simple meridian (e.g. RA = 12h).