KAPPA: Correct bad-pixel flag in COPYBAD.
The output NDF bad-pixel flag should be a combination (OR) of whether
or not any bad values were copied from the reference NDF, and the
presence of existing bad values in the input NDF.  As coded it was
possible to inherit a false bad-pixel flag even if bad values had been
introduced during the copy.

This is what I should have done in ea3515667db571.  Not sure why I
didn't want the overhead to check the input NDF for bad values.