smurf: New command jsapaster to do the opposite of jsadicer
That is, jsapaster pastes together a set of JSA tiles into a single map,
still gridded on the JSA all-sky pixel grid. It is similar to kappa:paste
except that it recognises cases where the output mosaic would cross an
HPX discontinuity, and converts the output mosaic to an appropriate XPH
projection instead. It also differs from kappa:paste in that it will join
together any NDFs contained within the smurf extensions of the input
NDFs.