gaia: fix issue with graphics capture and XQuartz 2.7.7.
Had a report that capturing a 3d animation failed, with the don't obscure
warning message. On inspection this was true under Mavericks with XQuartz
2.7.7, and also true for the image display capture for offscreen parts, (the
initial report was for Mountain Lion). The problem is that a call to XGetImage
in tkimg is returning NULL. Only fix that seems to work is changing some
timing waits, so this is probably some caching/buffering change in
XQuartz. Seems like a hack...