smurf: Single-thread parts of makemap if it will avoid chunking
Commit e93e76879e727 causes a big expansion in the memory needed to rebin
the time-stream data into a map. The required memory is now proprotional
to the number of threads. For machines with lots of threads, and for big
maps, this means that the memory requirement may become so big as to need
chunking. Such chunking can be avoided, at the cost of slower execution,
by reverting to a single-threaded version of smf_rebinmap1.

This commit adds code to see if chunking could be avoided by using a
single thread in smf_rebinmap1, and if so does so. The full
multi-threaded version is still used if no chunking is needed anyway, or
if reverting to a single thread would not prevent chunking being needed.