vtk: Need to explicitly state we are using X11 OpenGL
For a while now OS X vtk has been attempting to use the native
Framework regardless of the Cocoa/X11 setting. It was working because
X11 OpenGL was picked up by default anyhow.

Not tested on Linux (where it was presumably working fine without
any fiddling) so not sure if this patch should only be enabled
on OS X.