pub: FellWalker -  Gaussclumps would do better with test gaussians
Referee says "In the test case of Gaussian mixtures (Figs. 7, 8), the
obvious method to compare FellWalker to is a Gaussian mixture model.
I am not asking for this to be implemented and tested, but I suggest
the author speculate on the advantages and disadvantages of FellWalker
compared to that. One obvious benefit of FellWalker is that it assumes
no fixed shape, although a particular algorithm may perform better for
Gaussian mixtures."