pub: FellWalker - differentiate more clear between different sets of test data
The referee says "Although in Sect. 3.1, the text says that all the
Gaussian sources have equal amplitude, that does not appear to be so,
by eye. Is that because of an improbable number of overlaps, does my
eye deceive me, or is the text wrong? Fig. 11, on the other hand,
does seem to have a set of equal, superimposed Gaussians."

In fact he was mis-interpreting the scope of the statement about
constant peak values.