gaia: add support for table.load.fits SAMP mtype
Added a new table.load.fits method to the GaiaSampAgent.  This
method is based on the existing image.load.fits and table.load.votable
methods.  The interface of the display_table_ protected method
has been altered to allow it to be called by GaiaUrlGet: it now takes
the table ID as the first argument, a type argument has been
added and it sets $window in cat_windows_ rather than returning
it.