figaro: Reorganize open logic in COPOBJ
HDF5 does not allow a READONLY file to be reopened as
READWRITE when the file is still open. This breaks COPOBJ
when copying an object from one part of a file to another
location in the same file.

To overcome this we now use DAT_ASSOC to see if the source
object exists (as before), then DAT_REF to get the full path
to the object and then DAT_CANCL to free locators associated
with the parameter.

Once we know the destination is okay we then reopen the
source locator using HDS_FIND.