latexsupport: fixup for underscores
For some reason underscores in non-tt type text where being replaced
with a small high dot (unicode character &#x2d9 I think).  I've fixed
this up by replacing that glyph with an underscore. Would be nice to
know why it was doing it though... (and hopefully nobody tries to use
that character).