kappa: Add new parameter ALIGNREF to wcsalign and wcsmosaic
This parameter is a boolean which if set true causes alignment to be
performed in the coordinate system defined by the reference NDF. By
default, alignment occurs in the coordinate system defined by the AST
attributes AlignSystem, AlignStdOfRest, AlignTimeScale, etc, all of which
default to fixed values (ICRS/Wavelength/MJD for AlignSystem,
Heliocentrix for AlignStdOfRest, etc).

So for instance if you have two images which both have (az,el) axes but
which were taken at different times, then by default they will be aligned
in ICRS, meaning that stars will be align properly in the result. If
ALIGNREF is set TRUE, then they will be aligned in (az,el) (the system of
the reference NDF), meaning that stars will not be aligned propely since
they will have moved in (az,el) due to the different times of the two
images.