GSD: fix up sun229_mapping.tex and its script.
This should now work in html output. It uses a landscape page and a
sllongtable, and uses p formatting in the tabular so that the html
table will line wrap.  sllongtable is a slightly hacked longtable that
works reasonably in tex4ht output. If longtable is ever configured
more sensibly for tex4ht it would be slighty better to use that
environment instead.