HDSTRACE: Start to merge changes for UNIX era.
Tim beat me to converting SUN/102 to the new scheme.  It had been on my
desk for over two weeks after seeing that it was in LaTex 2.09 and had not
been converted to UNIX and hypertext era.  The first attempt to go to
LaTeX2e failed, but using the subsequently released new shorter prologue
looked a better approach (and so it proved as it built).