dtask: change a/ilink commands so that they compile dtask_main rather than use an installed dtask_main.o.
The signature of the main program can change with new compiler versions
so this is a most compatible option.

Triggered by problems with gfortran 4.5+ being incompatible with 4.4,
and yet another issue with 4.9 not returning the task name. Recompilation
of the main program routine resolved these issues and the optimisation
of having a single .o file is slight. Also allows use to remove the need
for dtask_main.txt as the proper source code is now present.