rv: Update SUN/78 to new latex
Also replace the PS figure (using \special commands) with PNG
since it was a screen grab to begin with and this prevents tex2ht
converting the PDF figure to PNG itself and messing it up.

The tabular appendix was not translating well so I converted it
to a normal terminalv.