dipso: SUN/50 in new latex
Rewrote the \dipcom command so that it used \subsubsection rather
than \item and also included a \label. Without this none of the
internal references to the commands in the document were
resolving. Now they work and also are clickable links in HTML.

The \htmlref usage to internal commands really needs to be replaced
with a macro to allow consistent use of the COM: prefix.