smurf: Report an error if a detector axis section is supplied to timesort.
timesort expects the lower bound of the detector axis to be 1, and the
upper bound to be equal to the number of detectors listed in the ACIS
extension. An error is now reported if this is not the case (i.e. a NDF
section is supplied as input). Previously, it ran but produced bad
results.