ast: Initial attempt at new latex for SUN/211 and SUN/210
addlinks is modified to prevent insertion of hyperlinks into
the list of figures and to also recognize that terminalv is
a verbatim block.

\sstlabel{} is still present but is disabled. This causes some
issues for internal references that are assuming that the label
is the normalized form and not the version given to \sstroutine.

For example "Top(axis)" differs in the routine name does not
match a label of "Topaxis". These need to be fixed.