latexsupport: Add definitions for SC/20.
It's not clear which packages, SUNs, and applications are sufficiently
common to warrant being in the master list.  The main packages
and their SUNs seems reasonable.  The choice of tasks is more fuzzy.
I added some more specialised tasks to SC/20.  Unless we decide to
include all or most, I would suggest removing things like \configecho
as it's probably only referenced in the one manual (SC/21).