docs: Explain hybrid and improve instructions for merging.
Add a brief subsection on what hybrid means.  Not sure if the location
for the Hybrid introduction is optimal.  Is it an observing mode?

Link to this when discussing merging of hybrid subsystems.

The merger instructions omitted correction of the DC offset and
the NDFCOPY commands were incomplete.  It could do with another,
less sleepy, pair of eyes to check the commands are now correct.