docs: Second go at the figures for SC/20.
The build is generating errors but completes.  Some graphics are
missing from the HTML.  I'm committng now as I have other SC/20
changes to commit too.

Early there are two sets of

  Error: PDF file is damaged - attempting to reconstruct xref table...
  Error: Couldn't find trailer dictionary
  Error: Couldn't read xref table

but it doesn't say which PDF.

At the end there are three sets of messages like this.

  System call: convert sc20_plotclumps2b.pdf sc20_plotclumps2b.png
  convert: unable to open image `sc20_plotclumps2b.pdf': No such file or directory @ blob.c/OpenBlob/2480.
  convert: missing an image filename `sc20_plotclumps2b.png' @ convert.c/ConvertImageCommand/2800.
  --- Warning --- System return: 256

Likewise for sc20_gaia-3dvolume2.png and sc20_clinplot.png.

I ran this command

   ps2pdf -DEPSCrop <blah>.eps

to make the conversion from EPS to PDF.