docs: Replace graphics macros with full markup in SC/20.
star2html looks for the \includegraphics but it isn't present verbatim
if the compact \myfig and \myfigduo macros are called.  Doing it
longhand avoids messiness in Makefile.am.  The alternative would be to
make star2html recognise the macros.

The macros had the centering misplaced before the caption.
I've put the label before the caption rather than inside.