docs: Reposition labels in figures.
My \myfig macro had these inside the caption, but it was easier to place the
labels after the caption using the learn mode in jed.  That was fine,
but then I saw Holly's explicit \includegraphics figures had the label
before the caption so I standardised on that, whereas I should have
corrected those to place the label after the caption to match mine.
As was it would link to the section heading not the caption.

Also fixed a typo in a label.