HTX: improved handling for ADS urls of obsolete documents.
There is now a single file (adsurl.txt) containing the ads ref for all
the starlink documents (obsolete and current). (This was taken from
the Starlink/adspubs repo on Github; it has slightly different
formatting from its source file 'star2bib.txt'). There is also a
separate file 'obsdoc.txt' that contains a list of all the obsolete
documents. When moving docs to obsolete, please update the list
'obsdoc.txt'. This change is to make it easier to keep this list
up-to-date.