pgplot: Add support for a few postscript fonts
This change is based on the code described at

http://members.home.nl/p.j.m.smulders/FLUX/HTML/psdriver.html

The environment variable PGPLOT_PS_FONT can now be set to one
of the following values:

Times
Courier
Helvetica
NewCentury
Zapf

to get the correspond font used in postscript output. The usual PGPLOT
font number (1-4) produces the following modification to the basic font
set by PGPLOT_PS_FONT:

1 - normal
2 - italic
3 - bold
4 - bold and italic

If PGPLOT_PS_FONT is not set, then the old PGPLOT fonts are used as before.
(note, this is changed from the behaviour of the patch as obtained from the
above web page).