star2html: no longer uses latex2html
This clears out the latex2html specific code in star2html.in,
removes the obsolete template files, and updates the build files.
It also removes various options that weren't being supported by
the TeX4HT build anyway; if anyone needs these we can come up with
a version for the TeX4HT build and add them back in.

MUD/152 (latex2html) was moved to obsolete.

SUN/199 is now *extremely* out of date. Most of it is discusing what
you can and can't do in .tex files that latex2html will be
converting. The equivalent for the new system will be in SUN/270 in
latexsupport, so there will be very little content left in SUN/199.