star2html: add splitlevel option for htlatex
This applies the hlatex command line option to split the document
up at different levels. This is to help the build of documents
that are intended to be displayed inside a java help browser. To
split at 'subsection' level e.g. you would usually call star2html
with the -splitlevel 3 option.