MacOs Build: first attempt at rpath fixup script
This commit adds in 'macos_postbuild_library_fixup.sh', which is
the initial attempt at a script to fixup the Mac OSX libraries in
a build so that all shared libraries shipped with starlink are referenced
through an @rpath, and all .dylibs and binaries have an appropriate rpath
set.

This script has currently only been slightly tested, and is not yet
ready for production use.  USE AT YOUR OWN RISK. Its logic is probably
a little odd in places.

If it works, it should a) allow DYLD_LIBRARY_PATH to not be set and
b) the mac build should be relocatable.