latexsupport: improve title page and cleanup code
This a) makes \stardocmanual and \stardocversion optional, b) adds an
optional parameter to \stardoctitle where users can give a short form
of the title for use in the page headings and the html page title. and
c) stops causing issues with linebreaks in the title.  In the process,
the code was generally tidied up and some no-longer required parts
were removed.