smurf: Add scan synchronous noise (SSN) model.
To use this model, include the following in your makemap config:

modelorder=(com,gai,ext,ssn,flt,ast,noi)

The SSN model attempts to remove scan-synchronous noise independently for
each individual bolometer.  For each bolometer, the residuals are first
binned according to their offset in azimuth from the tracking centre.
This produces a profile of the mean residual value as a function of
azimuth offset. The SSN value for each sample is then found by looking
up the approrpiate value in this profile.

Great effort is expended in various attempts to reduce the influence of
real astronomical signal on ths azimuth profile. Real astronomical signal
is characterised by the fact that it does not extend across the full
range of elevation offset.

Since the SSN seems to vary with time, each time stream is broken into
several sections and a separate model produced for each one. Linear
interpolation is performed between sections when estimating and removing
the SSN model.

Sadly, the resulting model is not generally very useful. It many cases it
seems to make little change to the map. In others, it actually makes
things worse. Why this is is not clear, but the following may be part of
the trouble:

1) The azimuth profiles can still be influenced by real astronomical signal,
despite all the efforts to avoid this, resulting in stripes being visible
across the whole the map, which run parallel to the elevation axis.

2) scan-synchronous noise (particularly within PONGS) seem not to be
always a simple function of azimuth offset, but also seem to depend on
the rate of change of azimuth.