latexsupport: improve sstroutine behaviour in tocs
By default, sstroutines will now have an entry in the main table of
contents, but not in the table of sections following each chapter or
section.  Instead, documents that want a list of the sstroutines
should use the command \sstminitoc{Title for List} at the appropriate
place. This command must be called just before a set of \sstroutine
entrys, and will list all of the sstroutine entries until it hits
another sectioning command.

In addition, users can separately turn off and turn back on some or
all of their sstroutines appearing in either of these table of
contents with the pairs of  commands: \sstnomaintoc & \sstmaintoc and
\sstnoextratoc & \sstextratoc