smurf: Ensure basic defaults reflect the values in dimmconfig.lis
The lists of parameter descriptions in SUN/258 and SC/21 are derived from
smurf_makemap.def. This includes the default values. But we've told users
to derived their configs from dimmconfig.lis. So what they are really
interested in is the default value provided by dimmconfig.lis, not the
default value provided by smurf_makemap.def. The simplest way to achieve
this is to change the defaults in smurf_makemap.def to match those in
dimmconfig.lis. This commit does that. But this introduces
a formal requirement that the defaults be the same dimmconfig.lis and
smurf_makemap.def. Which then raises the question of why we need
dimmconfig.lis at all.

But we cannot do away with dimmconfig.lis now as loads of users will
refer to it in their configs.

What we *could* maybe do, is to change dimmconfig.lis to be an empty
file. This would remove the need to implement any future changes of
default values in two separate places.