sc21: Update appendix containing list of parameter descriptions
The new file "selected_params.lis" contains the names of the parameters
that are to be included in SC21. The new file "selected_params.tex"
contains the tex for these parameters, created using the command:

%       ....applications/smurf/defaults/make_pardocs \
%               selected_params.lis \
%               selected_params.tex \
%               ../../../applications/smurf/defaults

(this command is included as a comment at the top of "selected_params.tex").

The file "sc21_selected_params.tex" is now a static file containing
sectioning commands and introductory text, prior to including the
descriptions themselves in selected_params.tex.

The list of parameters is no longer included in the main table of contents.

The parameter descriptions themselves have been updated to the current
text in smurf_makemap.def.

Some parameters have been removed fomr the selected list (WHITEN, PCA.*)