kappa: SUN/95 make section 9 look like rest of starlink docs.
The centered and quoted sections displaying how you would specify
different types of ndf sections were replaced with a terminalv
section. This is to create documents that are more uniform in
appearance and easier to maintain code.

By using the standard ~ character as well these can now be copied and
pasted safely into a terminal (previously the character used was
$\sim$ which appears as a unicode 'tilde operator' instead of the
nomral tilde, and therefore could not be cut and pasted into a
terminal on at least one computer.