asdt: Fix CAR plotting for most recent DS9 issue
Modify SpecTrans to avoid modifying the CRPIX value for CAR
projections when they do not need to be modified. The principle
is that the bulk of the array should be within the native longitude
range [-180,+180]. Prompted by bug report from Bill Joye "yet
another CAR issue" on 24-MAR-2015 (file CHIPASS_Equ.head in
ast_tester).