gaia: update news for 2015A release.
Update the on-line help for TAP. Includes an example of querying CADC for
SCUBA-2 observations.

Also include support for plotting STC regions, needed to move output format to
VOTable 1.2 and add support for xtype to manage that.

Added control to insert a simple query with the bounds of the image in place.