ast: astAddFrame now stores deep copies rather than references
Previously, this sort of code:

     AstFrameSet *fs = astRead( fitschan );
     AstPlot *plot = astPlot( fs, gbox, bbox, "System=Galactic" );

gave problems. The Plot is initialised using astAddFrame, which
previously stored references to the Frames within "fs". Setting
"System=galactic" on the new plot meant that the current Frame in "fs"
was also modified. Which is bad in itself. But what is worse is that this
broke the integrity of "fs" because the change was made via a Frame
pointer, rather than a FrameSet pointer.

This commit changes astAddFrame so that it stores deep copies, rather
than references, to the supplied Mapping and Frame. This means that the
Plot is now independent of "fs", so changes made to the plot have no
effect on "fs".

This is a change in behaviour. The new scheme is less likely to cause
unexpected behaviour than the old scheme, but it is just possible that
someone somewhere relies on the old behaviour ....