CONVERT: Respect FITS WCS headers when no WCS or AXIS component.
Some data sources, such as WFCAM, have put their WCS directly into the
FITS airlock, and do not have WCS or AXIS components.  So if the user
asks to propagate the FITS airlock in such cases, it should do so, and
not be clever devising different headers or removing headers.