Figaro: Use IAU/FITS standard units in code.
While Figaro mostly appears to be self-consistent in its use of units
such as A for Angstrom, Microns for um, Janskys for Jy (and milli-
and micro-fractions thereof), sec for second, it causes grief when
combined with other applications and AST that know about the standard.

So I have tried to use the standard strings for the UNITS components,
and be more flexible regarding the presence or not of the trailing s
in Microns or Angstroms.  Also tried to give a consistent order to
erg/s/cm**2/Angstrom.

Some string checking is messy because the Figaro character-handling
library does not have a string conversion to uppercase.

It's only been partly tested and there may be other occurrences as yet
unidentified that will need change to support IAU unit naming.

Also change the documentation to match the revised code.