ndf: Revert section axis order to wcs order if axes are at 45 degs
This change was prompted by the failure of the previous NDF section
scheme to handle sections of a JSA tile, all of which have sky axes that
are exactly 45 degrees away from the pixel axes.

The comma-separated fields within a "normal" NDF section string (i.e. one
that does not start with an asterisk), are assumed to be in pixel-axis
order. That is, the first field defines the extent on pixel axis 1, the
second field defines the extent on pixel axis 2, etc. It has to be this
way to handle cases where one or axes are specified by pixel bounds.

In cases where one or more axes are specified by WCS bounds, it is
therefore necessary to be able to associate each pixel axis (i.e. each
section field) with a specific WCS axis. Each pixel axis is associarted
with the WCS axis which is most nearly parallel to the pixel axis.

Example 1 - an image in which north is parallel to the pixel Y axis.

 % stats fred'(15:23:34~20,10:12:30~20)'

refers to a 20x20 pixel box centred on ra=15:23:34 dec=10:12:30. This
ordering does not depend on the order in which WCS axes are listed by
ndftrace. It is defined by the fact that the second pixel axis is
parallel to the Dec axis.

Example 2 - an image in which north is parallel to the pixel X axis.

 % stats fred'(15:23:34~20,10:12:30~20)'

refers to a 20x20 pixel box centred on dec=15:23:34 ra=10:12:30. This
ordering does not depend on the order in which WCS axes are listed by
ndftrace. It is defined by the fact that the first pixel axis is
parallel to the Dec axis.

Example 3- a JSA tile in which north is at exactly 45 degress to both
pixel axes.

 % stats fred'(15:23:34~20,10:12:30~20)'

The ordering of the fields cannot be determined by looking for parallel
axes. Previously, the ordering was indeterminate in such cases, leading
to unexpected interpretation of such sections. This commit identifies
such cases, and assumes that the fields are in the order in which WCS
axes are listed by ndftrace. For a JSA tile this is always (RA,Dec), so
the above section is now correctly interpreted as a 20x20 pixel box
centred on ra=15:23:34 dec=10:12:30.