smurf: Indicate we do not need sky positions to better than 0.05 arc-sec
This causes SlaMaps between SkyFrames to be replaced by UnitMaps during
simplification if the SlaMap produces a shift of less than 0.05
arc-seconds. This can clear a blockage that current prevents much
greater simplification taking place in the Mappings between pixel grids.

Enabling such simplification to take place is particularly important if
the total Mapping between pixel grids includes forward and inverse
PolyMaps describing SCUBA-2 distortion. A forward/inverse pair of
PolyMaps will cancel out if they are adjacent to each other in the
Mapping chain. Prior to this commit, an insignificant SlaMap could sit
between the two PolyMaps, preventing them from cancelling out. This is
bad since the two PolyMaps are only a (reasonably) accurate inverse pair
for positions that are within the SCUBA-2 focal plane. For positions that
are further out, applying the forward PolyMap followed by the inverse
PolyMap can leave large residual positional errors due to the inaccuracy
of the inversion. This commit (often) allows the two PolyMaps to cancel
out, resulting in a more accurate total transformation.

This is only an issue when doing an operation that is equivalent to
aligning time slices from different observations, such as is done within
the pol2cat script that removes backgrounds from "stare & spin" POL2
data.