KAPPA: Incomplete enhancement to BEAMFIT to allow for non-square pixels.
I'm making a preliminary commit before it's fully working as it's a
pressing upgrade, and I'm on leave for four days, to enable someone
else to take over, if necessary.  It should be safe as there is a new
SQUARE parameter to enable correction for non-square pixels, which
defaults to square pixels, giving the current behaviour.  Once the
scaling is working the file- and cursor-mode routines will need a
PIXSCR argument passing into the BF2_COM common block, as is done for
the interactive mode (kps1_bfint.f) subroutine.