kappa: beamfit - fix problems dealing with JSA HPX maps
1) Local variable REL was not initialised in kps1_bfft.f

2) Widths were taken as parallel to the pixel axes in kps1_bfcrf.f,
rather than parallel to the ellipse axes.

3) The small dynamic range of pixel coords (caused by huge pixel origins
in JSA maps) led to numerical problems in the minimisation algorithm. To
get round this, beamfit.f now temporarily shift the pixel origin to the
middle of the array.

In adddition, the uncertainties in the WCS frame are now found (in
kps1_bfcrf.f) by doing a statistical simulation involving 1000 tests,
perturbing the pixel positions by the associated pixel errors each time.
The WCS errors are then found by looking at the statistics of the
transformed, perturbed pixel positions. This does not seem to have a
significant effect on the timing of beamfit (I saw a slow down of about
2% - well within the noise of the timing operation).