kappa: Improve calculation of WCS errors in beamfit
kps1_bfcrf.f does a numerical simulation to calculate the WCS errors from
the pixel errors. The variances of the WCS values wwere previously found
using the usual "mean of the squares minus square of the mean" approach.
But this was giving useless values because it involves taking the
differences of two very large and very nearly equal values. So instead
use Welford's method which is much better conditioned. For instance, see

http://www.johndcook.com/blog/standard_deviation/